requirements.txt
slack-bolt>=1.18.0
apscheduler>=3.10.0
pytz>=2024.1
pyyaml>=6.0
httpx>=0.27.0
python-dotenv>=1.0.0
flask>=3.0.0
